deploy/requirements.txt
transformers==4.37.2
accelerate==0.24.1
tensorflow==2.15.0
tensorboardX==2.6.2.2
python-dotenv==1.0.0
scikit-learn==1.2.1
tqdm==4.65.0
Cython==0.29.33
matplotlib==3.7.0
numpy
pycocotools==2.0.6
Pillow==9.4.0
editdistance==0.6.2
scikit-image
scipy==1.10.1
PyMuPDF==1.21.1